+++ Asteroids Reimagined/PowerUp.h
@@ -35,8 +35,8 @@ public:
 	bool Initialize(Utilities* utilities);
 	bool BeginRun();
 
-	void Update(float deltaTime);
-	void FixedUpdate(float deltaTime);
+	void Update(double deltaTime);
+	void FixedUpdate(double deltaTime);
 	void Draw3D();
 
 	void Spawn(Vector3 position);
