+++ Knightmare/Entity.h
@@ -54,8 +54,8 @@ public:
 	virtual ~Entity();
 
 	virtual bool Initialize(Utilities* utilities);
-	virtual void Update(float deltaTime);
-	virtual void FixedUpdate(float deltaTime);
+	virtual void Update(double deltaTime);
+	virtual void FixedUpdate(double deltaTime);
 	virtual void Draw3D();
 
 	void X(float x);
